refactor: InitialCardNumberState 타입 하위에서 정의로 변경
